Fixed bandplot so that endpoint sends dollars instead of cents

+++ backend/src/api/types.rs
@@ -144,9 +144,9 @@ pub struct PercentileBandQuery {
 #[derive(Debug, Serialize)]
 pub struct PercentileDataPoint {
     pub block_number: u64,
-    pub percentile_25_cents: u64,
-    pub median_cents: u64,
-    pub percentile_75_cents: u64
+    pub percentile_25_cents: f64,
+    pub median_cents: f64,
+    pub percentile_75_cents: f64
 }
 
 #[derive(Debug, Serialize)]
